+++ admin/views/layout-footer.php
@@ -48,6 +48,7 @@ if (!isset($kernel)) {
 <script src="<?= esc_url(admin_asset_url('js/quick-edit.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/update-upload.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/update-progress.js')) ?>" defer></script>
+<script src="<?= esc_url(admin_asset_url('js/update-continue.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/select-all.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/confirm-submit.js')) ?>" defer></script>
 <script src="<?= esc_url(admin_asset_url('js/auto-submit.js')) ?>" defer></script>